change type

+++ data_loader_baxter.py
@@ -138,7 +138,7 @@ def load_test_dataset(env_names, data_path, pcd_path, importer, NP=100, min_leng
 	obs = load_normalized_dataset(env_names, pcd_path, importer)
 	obs = obs.reshape(len(obs),-1,3)
 	obs = importer.pointcloud_to_voxel(obs,
-			voxel_size=(32,32,32), padding_size=(32,32,32)).reshape(len(obs),1,32,32,32)
+			voxel_size=(32,32,32), padding_size=(32,32,32)).reshape(len(obs),1,32,32,32).astype(np.float32)
 
 	### obtain path length data ###
 	# paths_file = 'trainEnvironments_testPaths_GoalsCorrect_RRTSTAR_trainEnv_4.pkl'
